udpates

+++ pythonForPi/botPeripherals.py
@@ -126,7 +126,7 @@ class remoteControl(teensy):
         # CONTROLMODE,VEL,THETA,PHI
         # all values are relative to the forward direction of REx
         motionControllerMode= self.query('fcs').split(',')
-        return (float(motionControllerMode[0]),
+        return (int(motionControllerMode[0]),
                 float(motionControllerMode[1]),
                 float(motionControllerMode[2]),
                 float(motionControllerMode[4]))
